Squached some bugs

+++ app/Http/Controllers/CategoryController.php
@@ -53,6 +53,14 @@ class CategoryController extends Controller
 
     public function destroy(Category $Category)
     {
+
+        if ($Category->products()->exists()) {
+            // Redirect back with an error message
+            return redirect()->route('categories.index')->with('category-error', "Category '{$Category->name}' cannot be deleted because it has associated products.");
+        }
+
+
+
         $Category->delete();
         return redirect()->route('categories.index');
     }
